# Revert "use same storage class for standalone and distributed mode (fix #469)"