--- a/jubatus/server/server/classifier_serv.cpp
+++ b/jubatus/server/server/classifier_serv.cpp
@@ -66,7 +66,8 @@ struct classifier_serv_config {
 
 shared_ptr<core::storage::storage_base> make_model(
     const framework::server_argv& arg) {
-  return core::storage::storage_factory::create_storage("local_mixture");
+  return core::storage::storage_factory::create_storage(
+      (arg.is_standalone()) ? "local" : "local_mixture");
 }
 
 }  // namespace