--- a/jubatus/server/server/regression_serv.cpp
+++ b/jubatus/server/server/regression_serv.cpp
@@ -61,7 +61,8 @@ struct regression_serv_config {
 
 shared_ptr<core::storage::storage_base> make_model(
     const framework::server_argv& arg) {
-  return core::storage::storage_factory::create_storage("local_mixture");
+  return core::storage::storage_factory::create_storage(
+      (arg.is_standalone()) ? "local" : "local_mixture");
 }
 
 }  // namespace