This reverts commit 26362610725c42190b7dbcd2fb3377d467424fee. Refs #1190