@kmaehashi 2017-04-19
This PR must be merged after https://github.com/jubatus/jubatus_core/pull/359.

I run benchmarks over these patches and confirmed the improvement, using News20 jubatus-tutorial-python example (``time python tutorial.py > /dev/null``).

### Before: using local_storage_mixture (Jubatus 1.0.2), via RPC

Average runtime: 49.721 sec.

```
python tutorial.py > /dev/null  9.86s user 2.40s system 24% cpu 49.586 total
python tutorial.py > /dev/null  9.80s user 2.37s system 24% cpu 50.215 total
python tutorial.py > /dev/null  9.43s user 2.49s system 24% cpu 49.362 total
```

### After: using local_storage (#1193 + https://github.com/jubatus/jubatus_core/issues/359), via RPC

Average runtime: 41.975 sec.

```
python tutorial.py > /dev/null  9.80s user 2.34s system 29% cpu 41.746 total
python tutorial.py > /dev/null  9.82s user 2.35s system 28% cpu 42.269 total
python tutorial.py > /dev/null  9.52s user 2.49s system 28% cpu 41.910 total
```