@kmaehashi 2017-04-19
The only downside of this PR spec is that feature vectors which are not MIXed yet are ignored in standalone mode when loading models generated in distributed mode.  When you are trying to migrate from distributed to standalone, you need to save model after doing MIX.