@shiodat 2017-04-20
:100: 

I also tested with [this script](https://gist.github.com/shiodat/99470184ac92a3c42c8a88b29b2ef06d) and confirmed the improvement with these patches.

#### Before: using local_storage_mixture (Jubatus 1.0.2), via RPC and Embedded
```
$ python benchmark.py  
[server] batch_size: 1, train_time: 27.2937[s]
[server] batch_size: 10, train_time: 20.1504[s]
[server] batch_size: 100, train_time: 18.9459[s]
[embedded] batch_size: 1, train_time: 18.7774[s]
[embedded] batch_size: 10, train_time: 18.7828[s]
[embedded] batch_size: 100, train_time: 18.5782[s]
```

#### After: using local_storage (#1193 + jubatus/jubatus_core#359), via RPC and Embedded
```
$ python benchmark.py
[server] batch_size: 1, train_time: 22.8101[s]
[server] batch_size: 10, train_time: 15.9985[s]
[server] batch_size: 100, train_time: 14.6249[s]
[embedded] batch_size: 1, train_time: 14.3755[s]
[embedded] batch_size: 10, train_time: 12.8956[s]
[embedded] batch_size: 100, train_time: 13.1598[s]
```

We have to announce the downside of this PR by website and blog.