@kmaehashi 2017-04-20
> The only downside of this PR spec is that feature vectors which are not MIXed yet are ignored in standalone mode when loading models generated in distributed mode. When you are trying to migrate from distributed to standalone, you need to save model after doing MIX.

Ah, one more thing to note: when loading models saved in standalone mode of Jubatus 1.0.0-1.0.2 to Jubatus 1.0.3, it behaves as if they are saved in distributed mode, as 1.0.0-1.0.2 is using ``local_storage_mixture``.